change type of return variable to same type as return type

+++ 0x12-more_singly_linked_lists/101-print_listint_safe.c
@@ -9,7 +9,7 @@
  */
 size_t print_listint_safe(const listint_t *head)
 {
-	unsigned int count;
+	size_t count;
 	unsigned int i;
 	const listint_t *temp1;
 	const listint_t *temp2;
@@ -23,7 +23,7 @@ size_t print_listint_safe(const listint_t *head)
 		printf("[%p] %d\n", (void *)temp1, temp1->n);
 		count++;
 		temp1 = temp1->next;
-		temp2 = head; 
+		temp2 = head;
 		for (i = 0; i < count; i++)
 		{
 			if (temp1 == temp2)
